Don't use view composer but tag for cascade
If we are generating the cascade with the view composer, we are missing out on data that will be processed once Statamic parses the view. This is the case for any tag for instance.

I ran into this when trying to get the paginator from a collection tag with `Blink::get('tag-paginator')`. In the view composer, this would still return null as it has not been run yet.

Refacotring to a simple tag that returns the cascade resolves this issue. Only "problem" not is that the view cascade is not as pretty and poluted with `seo_` values. But this is worth the trade off. We should probably not fight how Statamic handles the cascade anyways.