Fix issue with taxonomy url not working anymore
To get the taxonomy’s url, we are setting the current site in the TaxonomySitemapUrl. This issue was caused because we used methods in the show.blade.php view instead of casting the values to an array beforehand. So when calling the methods in the view, it would register the wrong site.