Evaluate the conditions by model ID instead of URL
Had to change the `id` method on the `SeoDefaultSet` class and pass the ID to the view. Now we can get all `DefaultsData` in the `ConditionsController` without having to parse the URL. The only issue still to be solved is that we have to get the correct locale, which is not yet possible because we are not on the `Statamic::isCpRoute`.