Remove hreflang of localization with no route
We could have a german localization of an entry. But the german site doesn’t have a route defined in the collection settings. This change makes sure, that we remove all the hrelflangs of localizations with no route, so that we don’t output a non-existing link on the frontend.