Don’t consider disabled features anymore
This implementation is fighting a core fundamental of Statamic’s data flow as outlines here: https://github.com/statamic/ideas/issues/931

Let’s not fight this issue as this implementation would only work with a few custom conditions anyways.

We could probably find a different way to remove the data of disabled fields in a listener. For instance when saving the social media defaults.