Only add fields of enabled features
This ensures that we won’t add any data to the cascade if a field doesn’t have a fieldtype and doesn’t know how to augment itself. This typically happens when a field is not part of the blueprint anymore, but still has some remaining data in the content.

We also introduce isDisabledFeature again, which lets us determine, if a field should be added to the blueprint or not. This is a more elegant way than using the previous approach in the fieldtype’s augment method.